Modified all EventRequest pages
All EventRequest pages are panelized and re-styled. The index page
displays each requested event's name, details, time, and actions.

When viewing or approving an event, you'll notice an error message under
the "Contact Email" and "Contact Phone" headings. That indicates that we
need to create variables holding the contact email address and phone
number of the person requesting the event.